Web/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
numpy